pyproject.toml
[project]
name = "dotenv-platform-test"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jerome Hodges", email = "jh4@jainfamilyinstitute.org" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "tox>=4.23.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dotenv_platform_test"]
